+++ packages/sit-onyx/src/index.ts
@@ -20,6 +20,9 @@ export * from "./components/OnyxCheckbox/types";
 export { default as OnyxCheckboxGroup } from "./components/OnyxCheckboxGroup/OnyxCheckboxGroup.vue";
 export * from "./components/OnyxCheckboxGroup/types";
 
+export { default as OnyxDialog } from "./components/OnyxDialog/OnyxDialog.vue";
+export * from "./components/OnyxDialog/types";
+
 export { default as OnyxEmpty } from "./components/OnyxEmpty/OnyxEmpty.vue";
 
 export { default as OnyxFlyoutMenu } from "./components/OnyxFlyoutMenu/OnyxFlyoutMenu.vue";
@@ -32,6 +35,9 @@ export * from "./components/OnyxIcon/types";
 
 export { default as OnyxIconButton } from "./components/OnyxIconButton/OnyxIconButton.vue";
 export * from "./components/OnyxIconButton/types";
+
+export { default as OnyxInfoTooltip } from "./components/OnyxInfoTooltip/OnyxInfoTooltip.vue";
+export * from "./components/OnyxInfoTooltip/types";
 
 export { default as OnyxInput } from "./components/OnyxInput/OnyxInput.vue";
 export * from "./components/OnyxInput/types";