+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,23 @@
 # sit-onyx
 
+## 1.0.0-alpha.131
+
+### Major Changes
+
+- 1baef56: refactor(OnyxTable): rename property grid to withVerticalBorders
+
+## 1.0.0-alpha.130
+
+### Minor Changes
+
+- 37ee3fd: feat: add `OnyxDialog` component
+
+## 1.0.0-alpha.129
+
+### Minor Changes
+
+- 70c7f93: feat(OnyxInfoTooltip): export info tooltip
+
 ## 1.0.0-alpha.128
 
 ### Patch Changes